refactor pageHeader project location, add PortfolioWalletList widget
