+++ src/2-pages/auth/ui/SignupPage.tsx
@@ -1,7 +1,8 @@
 import {useEffect, useState} from 'react';
 import {AuthLayout} from './AuthLayout.tsx';
+import {PageHeader} from '@pages/ui';
 import {AuthModel} from '@entities/auth';
-import {Button, PageHeader, StatusPopup, TextField} from '@shared/ui';
+import {Button, StatusPopup, TextField} from '@shared/ui';
 import {cn, useKeyClick} from '@shared/lib';
 import {APP_PATH, APP_TEXT} from '@shared/constants';
