+++ src/2-pages/goal/routes/goal.routes.tsx
@@ -1,6 +1,6 @@
 import {APP_PATH} from '@shared/constants';
 import {PrivateRoute} from '@shared/lib';
-import {FullScreenPageLayout, AppLayout} from '@pages/ui';
+import {PageFullScreenLayout, AppLayout} from '@pages/ui';
 import {GoalCreatePage} from '../ui/GoalCreatePage.tsx';
 import {GoalDetailsPage} from '../ui/GoalDetailsPage.tsx';
 import {GoalListPage} from '../ui/GoalListPage.tsx';
@@ -71,7 +71,7 @@ const goalDetailsTransferRoute = {
 export const goalRoutes = [
 	{element: <AppLayout />, children: [goalListRoute]},
 	{
-		element: <FullScreenPageLayout />,
+		element: <PageFullScreenLayout />,
 		children: [
 			goalDetailsRoute,
 			goalCreateRoute,
