+++ src/2-pages/goal/ui/GoalDetailsPage.tsx
@@ -1,10 +1,11 @@
-import {PageWidgetsWrapper} from '@pages/ui';
+import {PageHeader, PageWidgetsWrapper} from '@pages/ui';
 import {GoalImage, GoalProgress, GoalTransactions} from '@widgets/goal';
+import {APP_PATH} from '@shared/constants';
 
 export function GoalDetailsPage() {
 	return (
 		<>
-			<GoalImage />
+			<GoalImage pageHeader={<PageHeader backPath={APP_PATH.goal.list} className='flex-grow' />} />
 			<PageWidgetsWrapper withTopSpace>
 				<GoalProgress />
 				<GoalTransactions />
