+++ src/2-pages/goal/ui/GoalTransactionsPage.tsx
@@ -1,7 +1,6 @@
 import {useParams} from 'react-router-dom';
-import {PageWidgetsWrapper} from '@pages/ui';
+import {PageHeader, PageWidgetsWrapper} from '@pages/ui';
 import {GoalTransactionsHistory} from '@widgets/goal';
-import {PageHeader} from '@shared/ui';
 import {APP_PATH, APP_TEXT} from '@shared/constants';
 
 export function GoalTransactionsPage() {
