+++ src/2-pages/ui/PageFullScreenLayout.tsx
@@ -1,6 +1,6 @@
 import {Outlet} from 'react-router-dom';
 
-export function FullScreenPageLayout() {
+export function PageFullScreenLayout() {
 	return (
 		<div role='screen-layout' className='mx-auto flex min-h-screen max-w-[33rem] flex-col'>
 			<Outlet />
