+++ src/2-pages/ui/index.ts
@@ -1,6 +1,8 @@
-export {FullScreenPageLayout} from './FullScreenPageLayout.tsx';
 export {AppLayout} from './AppLayout.tsx';
+
+export {PageFullScreenLayout} from './PageFullScreenLayout.tsx';
 export {PageWidgetsWrapper} from './PageWidgetsWrapper.tsx';
+export {PageHeader} from './page-header/ui/PageHeader.tsx';
 
 export {TransferPage} from './money-action-page/ui/TransferPage.tsx';
 export {FundWithdrawPage} from './money-action-page/ui/FundWithdrawPage.tsx';
