+++ src/3-widgets/goal/goal-image/ui/GoalImage.tsx
@@ -1,11 +1,12 @@
+import {ReactElement} from 'react';
 import {useParams} from 'react-router-dom';
 import {getButtonConfigs} from '../config/GoalImage.config.tsx';
 import {GoalModel} from '@entities/goal';
-import {Button, LoadingWrapper, PageHeader} from '@shared/ui';
-import {APP_PATH, APP_TEXT, CURRENCY_SYMBOL} from '@shared/constants';
+import {Button, LoadingWrapper} from '@shared/ui';
+import {APP_TEXT, CURRENCY_SYMBOL} from '@shared/constants';
 import {TextHelpers} from '@shared/lib';
 
-export function GoalImage() {
+export function GoalImage({pageHeader}: {pageHeader: ReactElement}) {
 	const {id} = useParams();
 	const {goalDetails, isGoalDetailsLoading} = GoalModel.useItemDetails({id});
 
@@ -15,7 +16,7 @@ export function GoalImage() {
 
 	return (
 		<div className='flex h-[310px] flex-col bg-secondary-grey'>
-			<PageHeader backPath={APP_PATH.goal.list} className='flex-grow' />
+			{pageHeader}
 
 			<div className='flex flex-col gap-2 px-4 py-2'>
 				<LoadingWrapper isLoading={isLoading} className='mb-5 h-6 w-14'>
