+++ src/3-widgets/portfolio/index.ts
@@ -1,4 +1,5 @@
 export * from './portfolio-management/ui/PortfolioManagement.tsx';
 export * from './portfolio-wallets/ui/PortfolioWallets.tsx';
+export * from './portfolio-wallet-list/ui/PortfolioWalletList.tsx';
 
 export * from './util/portfolio.constants.ts';
